+++ src/base.js
@@ -576,7 +576,7 @@ layerUI.init = function init(settings) {
  *
  * @type {String}
  */
-layerUI.version = '0.10.0';
+layerUI.version = '0.10.1';
 
 const clientVersions = Layer.Client.version.split('.').map(value => Number(value));
 if (clientVersions[0] !== 3 && Layer.Client.version !== '3.1.1') {