+++ src/components/conversation-list-panel/layer-conversations-list/layer-conversations-list.js
@@ -45,6 +45,7 @@
  * @extends layerUI.components.Component
  * @mixin layerUI.mixins.List
  * @mixin layerUI.mixins.MainComponent
+ * @mixin layerUI.mixins.ListSelection
  */
 import Layer from 'layer-websdk';
 import { registerComponent } from '../../../components/component';